Implement staging and runtime hooks
The buildpack team tried to implement pre- and post-staging hooks as an
experiment:

https://www.pivotaltracker.com/n/projects/1042066/stories/96592310

That approach didn't really work out because it requires the
functionality to be implemented in each buildpack.

This change implements the change directly inside the buildpackrunner,
making the functionality available even to buildpacks not based on
"compile-extensions".

The run-hook.sh adds similar functionality to the launcher.  If the
pre-hook.sh installs additional OS packages, then similar packages
might be needed before launching because the OS packages are not
persisted inside the droplet.