BAL works on windows2012R2
There are two main changes that need to be made in order for the BAL to
work on windows2012R2:

  1. Because of the way "containers" are implemented (directories in
  c:\containerizer), we need to prepend the working directory of the
  process running builder.exe to all of the builder config paths

  2. We can't assume that tar.exe is on the PATH. Instead, we will
  assume that it is next to the builder.exe file (i.e. in the same
  directory as all the other lifecycle binaries)

There are now three ways to build the builder:

  1. For linux
  2. GOOS=windows -tags=windows2012R2 // for windows2012R2
  3. GOOS=windows                     // for windows2016

When running the builder tests for windows2012R2, the environment
variable TAR_URL must be set. The test suite will download this file and
place it next to the builder.exe built by gexec.Build

The tar.exe files that will be included in diego-release can be found
here: https://greenhouse.ci.cf-app.com/teams/main/pipelines/tar/resources/s3-bucket

No changes are required to the launcher.

[#151173124]

Signed-off-by: Sam Smith <sesmith177@gmail.com>