Use entrypoint prefix, if present
Under normal circumstances `bash -c` is called, this commit changes that
by inserting the value specified by `config: {entrypoint_prefix: XXX}`
in the `staging_info.yml`.

This change ensures the entire start command is passed to the entrypoint
prefix. Without this change, a compound start comand such as:

> `echo foo; rackup`

would be tokenized by `bash` such that only `echo foo` is passed to the
entrypoint prefix, and `rackup` would be invoked without the prefix.

NOTE: `entrypoint_prefix` functionality is *not* supported on Windows.

Use DeaStagingInfo instead of a separate private stagingInfo struct

Signed-off-by: Aram Price <pricear@vmware.com>

Signed-off-by: David Sabeti <sabetid@vmware.com>